Uses more descriptive type names in test

+++ tests/testMemory.cpp
@@ -12,11 +12,11 @@ protected:
     RTTBuffer buf{RTTBuffer::create<T>(DEFAULT_CAPACITY)};
 };
 
-class IntBufFixture : public RTTBufFixture<int> {
+class IntBufFixture : public RTTBufFixture<int32_t> {
 protected:
     void iota(size_t size = DEFAULT_CAPACITY) {
-        for (int i = 0; i < size; ++i) {
-            buf.emplaceBack<int>(i);
+        for (int32_t i = 0; i < size; ++i) {
+            buf.emplaceBack<int32_t>(i);
         }
     }
 };
@@ -25,8 +25,8 @@ namespace ConstructorAssignmentTests {
     STEST_F(IntBufFixture, Construct) { }
 
     void checkBufferEquality(const RTTBuffer& lhs, const RTTBuffer& rhs) {
-        for (int i = 0; i < lhs.size(); ++i) {
-            EXPECT_EQ(lhs.at<int>(i), rhs.at<int>(i));
+        for (int32_t i = 0; i < lhs.size(); ++i) {
+            EXPECT_EQ(lhs.at<int32_t>(i), rhs.at<int32_t>(i));
         }
         EXPECT_EQ(lhs.size(), rhs.size());
         EXPECT_EQ(lhs.capacity(), rhs.capacity());
@@ -49,8 +49,8 @@ namespace ConstructorAssignmentTests {
 
     void checkBufferMoved(const RTTBuffer& oldBuffer, const RTTBuffer& newBuffer, size_t oldSize, size_t oldCapacity, const std::type_info& oldType, size_t oldElementSize) {
         // Verify the new buffer is correct.
-        for (int i = 0; i < newBuffer.size(); ++i) {
-            EXPECT_EQ(newBuffer.at<int>(i), i);
+        for (int32_t i = 0; i < newBuffer.size(); ++i) {
+            EXPECT_EQ(newBuffer.at<int32_t>(i), i);
         }
         EXPECT_EQ(newBuffer.size(), oldSize);
         EXPECT_EQ(newBuffer.capacity(), oldCapacity);
@@ -60,7 +60,7 @@ namespace ConstructorAssignmentTests {
         // Verify buf
         EXPECT_EQ(oldBuffer.size(), 0);
         EXPECT_EQ(oldBuffer.capacity(), 0);
-        EXPECT_EQ(oldBuffer.data<int>(), nullptr);
+        EXPECT_EQ(oldBuffer.data<int32_t>(), nullptr);
     }
 
     STEST_F(IntBufFixture, MoveConstruct) {
@@ -94,43 +94,43 @@ STEST_F(IntBufFixture, TypeChecks) {
 #endif
 
 STEST_F(IntBufFixture, BoundsChecks) {
-    EXPECT_THROWS(buf.at<int>(buf.size()), std::out_of_range);
+    EXPECT_THROWS(buf.at<int32_t>(buf.size()), std::out_of_range);
 }
 
 STEST_F(IntBufFixture, UncheckedAccessorDoesNotThrow) {
-    buf.atUnchecked<int>(buf.size());
+    buf.atUnchecked<int32_t>(buf.size());
 }
 
 STEST_F(IntBufFixture, PushBack) {
-    for (int i = 0; i < DEFAULT_CAPACITY; ++i) {
+    for (int32_t i = 0; i < DEFAULT_CAPACITY; ++i) {
         buf.pushBack(i);
     }
-    for (int i = 0; i < buf.size(); ++i) {
-        EXPECT_EQ(buf.at<int>(i), i);
+    for (int32_t i = 0; i < buf.size(); ++i) {
+        EXPECT_EQ(buf.at<int32_t>(i), i);
     }
-    EXPECT_EQ(buf.elementSize(), sizeof(int));
+    EXPECT_EQ(buf.elementSize(), sizeof(int32_t));
 }
 
 STEST_F(IntBufFixture, EmplaceBack) {
-    for (int i = 0; i < DEFAULT_CAPACITY; ++i) {
-        buf.emplaceBack<int>(i);
+    for (int32_t i = 0; i < DEFAULT_CAPACITY; ++i) {
+        buf.emplaceBack<int32_t>(i);
     }
-    for (int i = 0; i < buf.size(); ++i) {
-        EXPECT_EQ(buf.at<int>(i), i);
+    for (int32_t i = 0; i < buf.size(); ++i) {
+        EXPECT_EQ(buf.at<int32_t>(i), i);
     }
-    EXPECT_EQ(buf.elementSize(), sizeof(int));
+    EXPECT_EQ(buf.elementSize(), sizeof(int32_t));
 }
 
 STEST(EmplaceBackDoesNotCauseExcessiveReallocation) {
     // To emplace 8 elements, only 3 reallocations should occur -> 0->2->4->8
-    RTTBuffer buf{RTTBuffer::create<int>(0)};
-    static constexpr int DESIRED_CAPACITY = 8;
+    RTTBuffer buf{RTTBuffer::create<int32_t>(0)};
+    static constexpr uint8_t DESIRED_CAPACITY = 8;
 
     size_t currentCapacity{buf.capacity()};
-    size_t numReallocations{0};
+    uint8_t numReallocations{0};
 
-    for (int i = 0; i < DESIRED_CAPACITY; ++i) {
-        buf.emplaceBack<int>(i);
+    for (int32_t i = 0; i < DESIRED_CAPACITY; ++i) {
+        buf.emplaceBack<int32_t>(i);
         if (buf.capacity() != currentCapacity) {
             currentCapacity = buf.capacity();
             ++numReallocations;
